Mudanças Gerais...

+++ Installer 357/Package/install.bat
@@ -568,19 +568,28 @@ echo %progress%%%
 
 if %ERROS% gtr 0 (
 echo - Ocorreram %ERROS% erros durante a instalação! > "%UpCore%\ProgressT.log"
+echo %date%-%time% Instalação Concluída! Erros: %ERROS% >> "%UpCore%\UpdateLog.txt"
 echo 100 > "%UpCore%\ProgressBar.log"
 echo 21 >> "%UpCore%\ProgressBar.log"
 ) else (
+if exist "registry.vbs" (
 echo - Atualizando Registro... > "%UpCore%\ProgressT.log"
 echo %date%-%time% Atualizando Registro... >> "%UpCore%\UpdateLog.txt"
 registry.vbs /code:%code% /destination:"%Destination%"
+echo %date%-%time% Instalação Concluída! >> "%UpCore%\UpdateLog.txt"
 echo - Instalação Concluída! > "%UpCore%\ProgressT.log"
 echo 100 > "%UpCore%\ProgressBar.log"
 echo 2 >> "%UpCore%\ProgressBar.log"
+) else (
+echo - Falha ao Atualizar Registro! > "%UpCore%\ProgressT.log"
+echo %date%-%time% Falha ao Atualizar Registro! >> "%UpCore%\UpdateLog.txt"
+echo %date%-%time% Arquivo de registro não encontrado! >> "%UpCore%\UpdateLog.txt"
+echo 100 > "%UpCore%\ProgressBar.log"
+echo 21 >> "%UpCore%\ProgressBar.log"
+)
 )
 CLS
 echo 100%%
-echo %date%-%time% Instalação Concluída! Erros: %ERROS% >> "%UpCore%\UpdateLog.txt"
 cd ..\
 del /Q /F /S UpInstalation\*
 rd /Q /S UpInstalation
